Modify ParentLocationCache to manually track its reverse lookups.
This is an optimization. Previously, ParentLocationCache.delete_by_value()
would iterate through all items in the dict. This could become very
expensive during the import of large courses. This commit creates a
reverse lookup dict so that we can avoid the iteration. Quick local
testing shows that this iteration could account for up to 1/3rd the
processing time when importing large courses.